Convert approval reducer to TS.
